perf: avoid query getChild inside list
Required data are owner, permissions, typename and id.
All these data are already available in cache since they are read
with the getChildren query.
Read required data from cache only.
Update children parent field during read of children field, when
the data of the both the folder and the children are already written in cache.

refs: FILES-477