test: update jest and update test and mocks
Update jest to use the fixed version of the restoreMocks util.
Fix all the mocks, removing the unnecessary ones and creating before each test
the other ones.

Remove the useNavigation mock
Fix the navigation hooks of the shell mocked module
Update the types of the local module mocks
Create the local module mock for the useCreateOptions hook
Update the types for the useCreateOptions hook
Move the global variables of the test env inside the jest setup file to avoid
having them mixed with the runtime ones
Remove the global variables which are not usefull anymore

refs: FILES-685